Add new command send-transaction-to-contract
In use for an option to call a contract function via the CLI, a new
command has been added. In extension to the deploy command it needs the
address of the already deployed contract. Furthermore it is searching
for a matching function in the ABI by name and argument length.
Duplicated function names with same argument length but different types,
can't be called, because of the argument type conversation.
Finally the transaction hash for a successful mined execution is printed
out.

Related to: #41
See also: trustlines-protocol/blockchain#130