Set `from` field for local transactions
When calling a web3.eth.Contract function, the `from` field does
not get set for the `eth_estimateGas` RPC call. This field is optional
by Parity. For the actual `eth_sendTransaction` call, this field is not
optional anymore and gets correctly set to the unlocked account of the
RPC node. This can lead to problems where a function access
`msg.sender`. This is especially the case for ownable contracts. Thereby
the estimation call will fail. As show case, the call web3 function call
does not fail, if the gas amount was specified manually. Here the
estimation will be skipped and the actual transaction works as expected.
As simple solution the `send_function_call_transaction` functions sets
the `from` field for the transaction options, if it does not send a raw
transaction .