Prefer web3.eth.defaultAccount for signing transactions
When setting the 'from' address for signing transactions, we did use
web3.eth.accounts[0] as default from address if no 'from' address had
been given.

With this commit, we will prefer web3.eth.defaultAccount if it's set.

This will allow us to run the relay server with a node, that doesn't
have an unlocked account or even has no account at all.

It would probably have been better to explicitly require the 'from'
address, but at the moment I don't want to make that change.